# Fix function generation reproducability