--- a/test/core/QCheck_expect_test.expected
+++ b/test/core/QCheck_expect_test.expected
@@ -1,5 +1,13 @@
 random seed: 1234
-50 7 4 2 2 2 2 2 2 2 2 2 2 3 3 3 3 3 3 3 3 3 3 2724675603984413065
+50 7 4 2 2 2 2 2 2 2 2 2 2 3 3 3 3 3 3 3 3 3 3 (6,fun,50) (0,fun,2) (7,fun,66) (54,fun,9) (2,fun,5) (90,fun,4) (1,fun,4) (37,fun,7) (7,fun,1) (8,fun,5)
+(6,fun,50) (0,fun,2) (7,fun,66) (54,fun,9) (2,fun,5) (90,fun,4) (1,fun,4) (37,fun,7) (7,fun,1) (8,fun,5)
+(6,fun,50) (0,fun,2) (7,fun,66) (54,fun,9) (2,fun,5) (90,fun,4) (1,fun,4) (37,fun,7) (7,fun,1) (8,fun,5)
+(6,fun,50) (0,fun,2) (7,fun,66) (54,fun,9) (2,fun,5) (90,fun,4) (1,fun,4) (37,fun,7) (7,fun,1) (8,fun,5)
+(6,fun,50) (0,fun,2) (7,fun,66) (54,fun,9) (2,fun,5) (90,fun,4) (1,fun,4) (37,fun,7) (7,fun,1) (8,fun,5)
+(6,fun,50) (0,fun,2) (7,fun,66) (54,fun,9) (2,fun,5) (90,fun,4) (1,fun,4) (37,fun,7) (7,fun,1) (8,fun,5)
+(6,fun,50) (0,fun,2) (7,fun,66) (54,fun,9) (2,fun,5) (90,fun,4) (1,fun,4) (37,fun,7) (7,fun,1) (8,fun,5)
+(6,fun,50) (0,fun,2) (7,fun,66) (54,fun,9) (2,fun,5) (90,fun,4) (1,fun,4) (37,fun,7) (7,fun,1) (8,fun,5)
+2724675603984413065
 1362337801992206533
 681168900996103267
 340584450498051634
@@ -465,19 +473,19 @@ Leaf 0
 
 --- Failure --------------------------------------------------------------------
 
-Test fail_pred_map_commute failed (127 shrink steps):
+Test fail_pred_map_commute failed (107 shrink steps):
 
-([3], {_ -> 0}, {3 -> false; _ -> true})
+([0], {_ -> -21}, {-21 -> true; _ -> false})
 
 --- Failure --------------------------------------------------------------------
 
 Test fail_pred_strings failed (1 shrink steps):
 
-{some random string -> true; _ -> false}
+{some other string -> false; _ -> true}
 
 --- Failure --------------------------------------------------------------------
 
-Test fold_left fold_right failed (25 shrink steps):
+Test fold_left fold_right failed (24 shrink steps):
 
 (0, [1], {(1, 0) -> 1; _ -> 0})
 
@@ -490,15 +498,15 @@ l=[1], fold_left=1, fold_right=0
 
 --- Failure --------------------------------------------------------------------
 
-Test fold_left fold_right uncurried failed (111 shrink steps):
+Test fold_left fold_right uncurried failed (97 shrink steps):
 
-({(5, 7) -> 0; _ -> 7}, 0, [5; 0])
+({(1, 7) -> 0; _ -> 7}, 0, [1; 0])
 
 --- Failure --------------------------------------------------------------------
 
-Test fold_left fold_right uncurried fun last failed (26 shrink steps):
+Test fold_left fold_right uncurried fun last failed (21 shrink steps):
 
-(0, [1], {(0, 1) -> 1; _ -> 0})
+(0, [1], {(0, 1) -> 0; _ -> 1})
 
 --- Failure --------------------------------------------------------------------
 
@@ -1203,7 +1211,7 @@ stats dist:
    4150517416584649600.. 4611686018427387903: #################                                               189
 ================================================================================
 1 warning(s)
-failure (57 tests failed, 1 tests errored, ran 121 tests)
+failure (57 tests failed, 1 tests errored, ran 129 tests)
 random seed: 153870556
 
 +++ Stats for int_dist_empty_bucket ++++++++++++++++++++++++++++++++++++++++++++++++++++++++