This PR fixes #236

QCheck and QCheck2 generates functions by inhabiting (and side-effecting) an underlying `Hashtbl`, packaged as `Poly_tbl`.
A new function application then triggers a generator call and adds the new binding to the underlying table.
Overall, since the generator call side-effects the state of the underlying `Random.State`, the number of function applications may then unknowingly affect the output of another generator, e.g., if the function generator is part of a tuple as the example in #236:
```ocaml
(triple small_int (fun1 Observable.int small_int) small_int)
```
This side-effecting clearly breaks (what we tend to think of as) independence of the individual tuple entry generators.

The PR fixes the function generator by `split`ting the `Random.State` and lets the later function applications side-effect the split copy instead, thus restoring (some form of) independence. We do so for both `QCheck` and `QCheck2`.

The `RS.split` is part of the `Random.State` interface of the forth-coming OCaml 5.0 release - but we add a "poor man's" `split` for backwards compatibility on <5.0. 

Finally the PR adds tests - with repeated reruns. Without the fix both of them gave rise to different prints when repeated. Both are fixed with the proposed patch.